+++ src/common/image_process_utils/imageproc.cpp
@@ -11,7 +11,7 @@ namespace imgproc {
         int dst_w, int dst_h,
         bool antialias)
     {
-        std::vector<float> dst(dst_w * dst_h);
+        std::vector<float> dst(static_cast<size_t>(dst_w) * dst_h);
         
         // PyTorch's exact scale calculation (align_corners=False for resize)
         const float scale_w = area_pixel_compute_scale(src_w, dst_w, false);
